修复下载swoole 源码 错误

+++ sapi/src/builder/extension/swoole.php
@@ -8,7 +8,7 @@ return function (Preprocessor $p) {
     $libiconv_prefix = ICONV_PREFIX;
     $ngtcp2_prefix = NGTCP2_PREFIX;
     $nghttp3_prefix = NGHTTP3_PREFIX;
-    $dependentLibraries = ['curl', 'openssl', 'cares', 'zlib', 'brotli', 'nghttp2', 'sqlite3', 'unix_odbc', 'pgsql', 'libzstd', 'ngtcp2', 'nghttp3'];
+    $dependentLibraries = ['curl', 'openssl', 'cares', 'zlib', 'brotli', 'nghttp2', 'sqlite3', 'unix_odbc', 'pgsql', 'libzstd', 'nghttp3'];
     $dependentExtensions = ['curl', 'openssl', 'sockets', 'mysqlnd', 'pdo'];
 
     $options[] = '--enable-swoole';
@@ -25,7 +25,6 @@ return function (Preprocessor $p) {
     $options[] = '--enable-brotli';
     $options[] = '--enable-zstd';
     $options[] = '--enable-swoole-stdext';
-    $options[] = '--with-ngtcp2-dir=' . $ngtcp2_prefix;
     $options[] = '--with-nghttp3-dir=' . $nghttp3_prefix;
 
     if ($p->isLinux() && $p->getInputOption('with-iouring')) {
@@ -58,6 +57,8 @@ return function (Preprocessor $p) {
         $shell .= <<<'EOF'
 
         SWOOLE_VERSION=$(awk 'NR==1{ print $1 }' "sapi/SWOOLE-VERSION.conf")
+        ORIGIN_SWOOLE_VERSION=${SWOOLE_VERSION}
+        SWOOLE_VERSION=$(echo "${SWOOLE_VERSION}" | sed 's/[^a-zA-Z0-9]/_/g')
         CURRENT_SWOOLE_VERSION=''
 
         if [ -f "ext/swoole/CMakeLists.txt" ] ;then
@@ -75,7 +76,9 @@ return function (Preprocessor $p) {
             test -d ext/swoole && rm -rf ext/swoole
             if [ ! -f ${WORKDIR}/pool/ext/swoole-${SWOOLE_VERSION}.tgz ] ;then
                 test -d /tmp/swoole && rm -rf /tmp/swoole
-                git clone -b "${SWOOLE_VERSION}" https://github.com/diyism/swoole-src.git /tmp/swoole
+                git clone -b "${ORIGIN_SWOOLE_VERSION}" https://github.com/diyism/swoole-src.git /tmp/swoole
+                status=$?
+                if [[ $status -ne 0 ]]; then { echo $status ; exit 1 ; } fi
                 cd  /tmp/swoole
                 rm -rf /tmp/swoole/.git/
                 tar -czvf ${WORKDIR}/pool/ext/swoole-${SWOOLE_VERSION}.tgz .
